add api folder for requests
